refactor : 높은 의존도로, 가독성을 해쳐서 BridgeMap.jav 클래스 삭제 및 BridgeGame으로 통합
- 의미상 BridgeGame에 있는게 맞는거 같기도 합니다
- 게임에 맵 정보와 이동 기록이 저장되는게 맞는 것 같다.
