+++ src/main/java/bridge/domain/BridgeGame.java
@@ -1,33 +1,47 @@
 package bridge.domain;
 
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 public final class BridgeGame {
-    private final BridgeMap bridgeMap;
+
+    public final static String UP_SIDE = "U";
+    public final static String DOWN_SIDE = "D";
     private static int gameTryCount;
+    private final List<String> moveLogs;
+    private final List<String> map;
+
+    public BridgeGame(List<String> map) {
+        this.map = map;
+        moveLogs = new ArrayList<>();
+    }
 
     static {
         gameTryCount = 1;
     }
 
-    public BridgeGame(BridgeMap bridgeMap) {
-        this.bridgeMap = bridgeMap;
-    }
-
     public boolean move(String direction) {
-        return bridgeMap.isCorrectWay(direction);
+        int nextPosition = moveLogs.size();
+        moveLogs.add(direction);
+        return map.get(nextPosition).equals(direction);
     }
 
-    public void retry() {
+    public void reset() {
         gameTryCount++;
+        moveLogs.clear();
+    }
+
+    public boolean isGameSuccess() {
+        return (moveLogs.size() == map.size())
+                && (moveLogs.toString().equals(map.toString()));
     }
 
     public int getGameTryCount() {
         return gameTryCount;
     }
 
-    public boolean isGameSuccess(List<String> map, List<String> moveLogs) {
-        return (moveLogs.size() == map.size())
-                && (moveLogs.toString().equals(map.toString()));
+    public List<String> getMoveLogs() {
+        return Collections.unmodifiableList(moveLogs);
     }
 }
