+++ src/main/java/bridge/dto/GameInformationDto.java
@@ -1,22 +1,19 @@
 package bridge.dto;
 
 import bridge.domain.BridgeGame;
-import bridge.domain.BridgeMap;
 
 import java.util.List;
 
 public final class GameInformationDto {
 
-    private final BridgeMap bridgeMap;
     private final BridgeGame bridgeGame;
 
-    public GameInformationDto(BridgeMap bridgeMap, BridgeGame bridgeGame) {
-        this.bridgeMap = bridgeMap;
+    public GameInformationDto(BridgeGame bridgeGame) {
         this.bridgeGame = bridgeGame;
     }
 
     public List<String> getMoveLogs() {
-        return bridgeMap.getMoveLogs();
+        return bridgeGame.getMoveLogs();
     }
 
     public int getGameTryCount() {
